+++ app.js
@@ -1,56 +1,7 @@
-const path = require("path");
-const {
-  bootstrap,
-  docker,
-} = require("@kaholo/plugin-library");
+const { bootstrap } = require("@kaholo/plugin-library");
 
-const {
-  pathExists,
-  exec,
-} = require("./helpers");
-
-const KAHOLO_PUPPETEER_IMAGE = "buildkite/puppeteer";
-
-async function runPuppeteerTest(params) {
-  const {
-    puppeteerJSFile,
-    workingDirectory,
-    environmentalVariables = {},
-  } = params;
-
-  const absoluteWorkingDir = path.resolve(workingDirectory);
-  if (!await pathExists(absoluteWorkingDir)) {
-    throw new Error(`Path ${workingDirectory} does not exist on agent!`);
-  }
-
-  const puppeteerCommand = `node ${puppeteerJSFile}`;
-  const dockerCommand = docker.buildDockerCommand({
-    image: KAHOLO_PUPPETEER_IMAGE,
-    command: docker.sanitizeCommand(puppeteerCommand),
-    workingDirectory: "/project",
-    environmentVariables: environmentalVariables,
-    additionalArguments: [
-      "-v",
-      `${absoluteWorkingDir}:/project`,
-    ],
-  });
-
-  const commandOutput = await exec(dockerCommand, {
-    env: {
-      ...process.env,
-      ...environmentalVariables,
-    },
-  });
-
-  if (commandOutput.stderr && !commandOutput.stdout) {
-    throw new Error(commandOutput.stderr);
-  } else if (commandOutput.stderr) {
-    console.error(commandOutput.stderr);
-  }
-
-  return commandOutput.stdout;
-}
+const puppeteerService = require("./puppeteer-service");
 
 module.exports = bootstrap({
-  runPuppeteerTest,
+  runPuppeteerTest: puppeteerService.runPuppeteerTest,
 });
